* nis/nss_nis/nis-hosts.c (internal_gethostbyname2_r): int -> size_t
	* nis/nss_nis/nis-network.c (_nss_nis_getnetbyname_r): Likewise.
	* nis/nss_nis/nis-alias.c (_nss_nis_getaliasbyname_r): Likewise.
	* nis/nis_table.c (__create_ib_request): Likewise.

+++ nis/nis_table.c
@@ -1,4 +1,4 @@
-/* Copyright (c) 1997, 1998, 1999 Free Software Foundation, Inc.
+/* Copyright (c) 1997, 1998, 1999, 2003 Free Software Foundation, Inc.
    This file is part of the GNU C Library.
    Contributed by Thorsten Kukuk <kukuk@suse.de>, 1997.
 
@@ -30,7 +30,7 @@ __create_ib_request (const_nis_name name, unsigned int flags)
   struct ib_request *ibreq = calloc (1, sizeof (ib_request));
   char buf[strlen (name) + 1];
   nis_attr *search_val = NULL;
-  int search_len = 0;
+  size_t search_len = 0;
   char *cptr;
   size_t size = 0;
